magixpromotion/requirements/test.txt
-r base.txt
pytest
pytest-django
pytest-cov
factory-boy
freezegun>=1.2
wagtail-factories>=4.1
